Haskellbook.Examples

Documentation

module Haskellbook.Examples.UserInfo

data GreetingMessage

Constructors

GreetingMessage

Fields

greetingSalutation :: Salutation

A Salutation, like "Hello"

greetingTo :: Name

Name of the person that should be greeted

greetingFrom :: [Name]

Names of the people who are sending the greeting

newtype Salutation

Constructors

Salutation

Fields

getSalutation :: String

newtype Name

Constructors

Name

Fields

getName :: String

excitingMessage :: String -> String

defaultMessage :: GreetingMessage

A default greeting message that isn't attributed to anyone: GreetingMessage { greetingSalutation = Salutation "Hello" , greetingTo = Name "Friend" , greetingFrom = [] }

formatMessage :: GreetingMessage -> String

A GreetingMessage contains all of the information needed to generate a greeting using formatMessage. You can get a default greeting without attribution from defaultMessage. This makes it convenient to use record update syntax to construct a new greeting:

>>> formatMessage defaultMessage { greetingFrom = [ Name "A Haskeller"] }
"Welcome Friendos, from: A Haskeller"

data SortedList

makeSortedList :: [Int] -> Maybe SortedList

minimum :: SortedList -> Int

the problems here ; empty list and list not being sorted