Haskellbook.Examples.Introduction.CreatingModules

Description

This module serves as an example of how you can create a module. This comment will be placed at the top of the generated documentation.

Documentation

newtype Name

Constructors

Name

Fields

getName :: String

newtype Salutation

Constructors

Salutation

Fields

getSalutation :: String

data GreetingMessage

Constructors

GreetingMessage

Fields

greetingSalutation :: Salutation

A Salutation, like "Hello"

greetingTo :: Name

Name of the person that should be greeted

greetingFrom :: [Name]

Names of the people who are sending the greeting

defaultMessage :: GreetingMessage

A default greeting message that isn't attributed to anyone: GreetingMessage { greetingSalutation = Salutation "Hello" , greetingTo = Name "Friend" , greetingFrom = [] }

formatMessage :: GreetingMessage -> String

A GreetingMessage contains all of the information needed to generate a greeting using formatMessage. You can get a default greeting without attribution from defaultMessage. This makes it convenient to use record update syntax to construct a new greeting:

>>> formatMessage defaultMessage { greetingFrom = [ Name "A Haskeller"] }
"Welcome Friendos, from: A Haskeller"

testMessage :: String

A test message that you can use to see how messages are formatted.

>>> testMessage
"Welcome Friendos, from your friends: test example, example2, and example3"

excitingMessage :: String -> String